Bug Fixes and IMprovements

+++ scm_from_coregsitered_functional.sh
@@ -18,10 +18,16 @@ scm_coregsitered_functional () {
       
       fslmaths ${1} -Tmean mean_mc_func_cannulas
 
-      echo "When fsleyes opens, create a mask to include cannulas in the functional data and save it as mask_mean_mc_func_cannulas.nii.gz"
-      cp mask_mean_mc_func.nii.gz mask_mean_mc_func_cannulas.nii.gz
-      fsleyes mean_mc_func_cannulas.nii.gz mask_mean_mc_func_cannulas.nii.gz
-      rm -f mean_mc_func_cannulas.nii.gz
+      
+      if [[ -f mask_mean_mc_func_cannulas.nii.gz ]]; then
+        echo "Mask file mask_mean_mc_func.nii.gz already exists. Using the existing mask."
+      else
+        echo "When fsleyes opens, create a mask to include cannulas in the functional data and save it as mask_mean_mc_func_cannulas.nii.gz"
+        cp mask_mean_mc_func.nii.gz mask_mean_mc_func_cannulas.nii.gz
+        fsleyes mean_mc_func_cannulas.nii.gz mask_mean_mc_func_cannulas.nii.gz
+        rm -f mean_mc_func_cannulas.nii.gz
+      fi
+      
 
       # Create masks to include cannulas from functional data
       fslmaths mc_func.nii.gz -mas mask_mean_mc_func_cannulas.nii.gz cleaned_with_cannula_mc_func.nii.gz
